pyproject.toml
[project]
name = "omi-integration-app-realtime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "uvicorn>=0.36.0",
    "stripe>=7.0.0",
    "python-dotenv>=1.0.0",
]
